Create GET and Post for Workouts
