+++ pages/_app.js
@@ -4,6 +4,7 @@ import GlobalStyles from "../components/GlobalStyles";
 function MyApp({ Component, pageProps }) {
   const [isFavourite, setIsFavourite] = useState([]);
   const [exercises, setExercises] = useState([]);
+  const [workouts, setWorkouts] = useState([]);
 
   function handleFav(id, event) {
     event.preventDefault();
@@ -19,8 +20,18 @@ function MyApp({ Component, pageProps }) {
     setExercises(exercisesList);
   }
 
+  async function getWorkouts() {
+    const response = await fetch("/api/workouts");
+    const workoutsList = await response.json();
+    setWorkouts(workoutsList);
+  }
+
+  function handleAddWorkout(newWorkout) {
+    setWorkouts([...workouts, newWorkout]);
+  }
   useEffect(() => {
     getExercises();
+    getWorkouts();
   }, []);
 
   return (
@@ -32,6 +43,7 @@ function MyApp({ Component, pageProps }) {
         onFav={handleFav}
         isFavourite={isFavourite}
         exercises={exercises}
+        onAddWorkout={handleAddWorkout}
         workouts={workouts}
       />
     </>
