+++ pages/workouts/index.js
@@ -2,25 +2,9 @@ import { Fragment } from "react";
 import Head from "next/head";
 import Nav from "../../components/Footer/Nav";
 import Workouts from "../../components/Main/Workouts/Workouts";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 
-const Workout = () => {
-  const [workouts, setWorkouts] = useState([]);
-
-  useEffect(() => {
-    getWorkouts();
-  }, []);
-
-  async function getWorkouts() {
-    const response = await fetch("/api/workouts");
-    const workoutsList = await response.json();
-    setWorkouts(workoutsList);
-  }
-
-  function handleAddWorkout(newWorkout) {
-    setWorkouts([...workouts, newWorkout]);
-  }
-
+const Workout = ({ workouts }) => {
   return (
     <Fragment>
       <Head>
@@ -30,7 +14,7 @@ const Workout = () => {
           content="width=device-width, initial-scale=1, maximum-scale=1, "
         />
       </Head>
-      <Workouts workouts={workouts} onAddWorkout={handleAddWorkout} />
+      <Workouts workouts={workouts} />
       <Nav />
     </Fragment>
   );
